+++ backend/gongnomok-common/src/main/java/site/gongnomok/common/management/dto/response/comment/ReportCommentDto.java
@@ -1,4 +1,4 @@
-package site.gongnomok.common.management.dto.response;
+package site.gongnomok.common.management.dto.response.comment;
 
 import lombok.*;
